Split the source upgrade task into brando.migrate54 and brando.migrate55
`mix brando.migrate54` grew to cover both the 0.53 to 0.54 rewrites and
everything added on `next` since, so a project that had already run the
0.54 version could not rerun it: the consumer-owned `brando.upgrade` task
and `scripts/sync_gettext.sh` it had been given no longer matched the
current templates and blocked the whole Igniter plan.

The rewrites now live in `Mix.Brando.Igniter.SourceUpgrade`, and two thin
tasks compose them by version:

- `brando.migrate54` keeps the 0.53 to 0.54 surface: Blueprint and
  LivePreview rewrites, the digest, font and repo_module fixes, and the
  Gettext update.
- `brando.migrate55` takes the 0.54 to 0.55 surface: listing component
  imports, the Req Swoosh client, the phoenix_live_view pin, Florist
  conversion, refreshing the gettext helper, and archiving the consumer
  upgrade task.

Every version of the consumer upgrade task Brando ever installed is kept
under priv/templates/brando.migrate/legacy_upgrade_tasks so the prepare
step recognizes the copy 0.54 shipped, and `mix brando.install` no longer
copies that task into new applications. The versioned `igniter.upgrade`
hook composes `brando.migrate55` for a 0.54 to 0.55 upgrade instead of
rejecting 0.55 outright.
