+++ lib/brando/migration/florist_config.ex
@@ -399,7 +399,7 @@ defmodule Brando.Migration.FloristConfig do
     database_variables = Enum.map_join(targets, ", ", &database_password_variable(&1.name))
 
     """
-    # Florist configuration generated by `mix brando.migrate54`.
+    # Florist configuration generated by `mix brando.migrate55`.
     # Source: legacy deployment.cfg + fabfile.py. Review every value before use.
     # Database passwords: #{database_variables}
     # SSH authentication uses your agent unless you add an environment-backed pass.
