+++ test/mix/brando/igniter/upgrade_test.exs
@@ -6,7 +6,11 @@ defmodule Mix.Brando.Igniter.UpgradeTest do
   @legacy_path "lib/mix/brando.upgrade.ex"
   @archive "priv/brando/legacy_tasks/brando.upgrade.ex.disabled"
 
-  defp legacy, do: File.read!(Application.app_dir(:brando, "priv/templates/brando.install/lib/mix/brando.upgrade.ex"))
+  @legacy_templates "priv/templates/brando.migrate/legacy_upgrade_tasks"
+
+  defp legacy(version \\ "0.55-dev") do
+    File.read!(Application.app_dir(:brando, [@legacy_templates, "brando.upgrade.#{version}.ex"]))
+  end
 
   test "recognized legacy source is archived with its comments before releasing the task name" do
     contents = "# My historical installer\n" <> legacy()
@@ -24,6 +28,22 @@ defmodule Mix.Brando.Igniter.UpgradeTest do
     assert rerun.issues == []
     assert rerun.rms == []
     Igniter.Test.assert_unchanged(rerun)
+  end
+
+  test "the task installed by Brando 0.54 is recognized and archived" do
+    result =
+      IgniterCase.phoenix_project(files: %{@legacy_path => legacy("0.54")})
+      |> Igniter.compose_task("brando.upgrade.prepare", [])
+
+    assert result.issues == []
+    Igniter.Test.assert_rms(result, @legacy_path)
+    assert IgniterCase.source(result, @archive) == legacy("0.54")
+  end
+
+  test "the installer no longer ships a consumer-owned upgrade task" do
+    refute Enum.any?(Mix.Brando.Install.Templates.manifest(), fn {_format, _source, target} ->
+             target == "lib/mix/brando.upgrade.ex"
+           end)
   end
 
   test "customized legacy tasks and archive collisions block removal even with yes" do
@@ -77,6 +97,22 @@ defmodule Mix.Brando.Igniter.UpgradeTest do
       result = IgniterCase.phoenix_project() |> Igniter.compose_task(Mix.Tasks.Brando.Upgrade, args)
       assert result.issues != []
       assert result.tasks == []
+      Igniter.Test.assert_unchanged(result)
+    end
+  end
+
+  test "a 0.54 to 0.55 upgrade composes the brando.migrate55 recipe before planning migrations" do
+    installed = Application.spec(:brando, :vsn) |> to_string() |> Version.parse!()
+    result = IgniterCase.phoenix_project() |> Igniter.compose_task(Mix.Tasks.Brando.Upgrade, ["0.54.0", "0.55.0-dev"])
+
+    if installed.minor >= 55 do
+      assert result.issues == []
+      assert Enum.any?(result.notices, &String.contains?(&1, "Brando 0.55 source migration prepared"))
+      assert Enum.any?(result.notices, &String.contains?(&1, "missing framework migrations"))
+    else
+      # Until the dependency itself is 0.55, the hook must refuse rather than
+      # apply a recipe for a version that is not loaded.
+      assert Enum.any?(result.issues, &String.contains?(&1, "cannot apply source upgrades for 0.55.0-dev"))
       Igniter.Test.assert_unchanged(result)
     end
   end
